updated sceneLoaderController to apply single responsibility principle
